requirements.txt
torch
transformers
datasets
evaluate
fastapi
uvicorn
pydantic
python-dotenv
python-jose[cryptography]
passlib
bcrypt<5.0
python-multipart
